% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{bf}
\alias{bf<-}
\alias{cells}
\alias{cells<-}
\alias{cf}
\alias{cf<-}
\alias{filename}
\alias{stages}
\alias{stages<-}
\alias{units}
\alias{units<-}
\alias{wells}
\alias{wells<-}
\alias{bf,Seahorse-method}
\alias{bf<-,Seahorse-method}
\alias{cells,Seahorse-method}
\alias{cells<-,Seahorse-method}
\alias{cf,Seahorse-method}
\alias{cf<-,Seahorse-method}
\alias{filename,Seahorse-method}
\alias{stages,Seahorse-method}
\alias{stages<-,Seahorse-method}
\alias{units,Seahorse-method}
\alias{units<-,Seahorse-method}
\alias{units,Herd-method}
\alias{units<-,Herd-method}
\alias{wells,Seahorse-method}
\alias{wells<-,Seahorse-method}
\title{Accessors for \code{seahorse} S4 objects}
\usage{
\S4method{bf}{Seahorse}(x)

\S4method{bf}{Seahorse}(x) <- value

\S4method{cells}{Seahorse}(x)

\S4method{cells}{Seahorse}(x) <- value

\S4method{cf}{Seahorse}(x)

\S4method{cf}{Seahorse}(x) <- value

\S4method{filename}{Seahorse}(x)

\S4method{stages}{Seahorse}(x)

\S4method{stages}{Seahorse}(x) <- value

\S4method{units}{Seahorse}(x)

\S4method{units}{Seahorse}(x) <- value

\S4method{units}{Herd}(x)

\S4method{units}{Herd}(x) <- value

\S4method{wells}{Seahorse}(x)

\S4method{wells}{Seahorse}(x) <- value
}
\arguments{
\item{x}{A \code{Seahorse} or \code{Herd} object.}

\item{value}{A replacement value.}
}
\value{
\code{bf(x)} returns the medium buffer factor.

\code{cells(x)} returns the normalization values used to adjust the rate
measurements.

\code{cf(x)} returns the carbon dioxide correction factor.

\code{filename(x)} returns the base filename for the \code{Seahorse}
experiment.

\code{stages(x)} returns the experimental design details associated with
each stage.

\code{units(x)} returns the character vector describing the normalization
units.

\code{units(x)} returns the character vector describing the normalization
units.

\code{wells(x)} returns the experimental design details associated with
each well.
}
\description{
These functions describe how to access and replace values in \link{Seahorse} and
\link{Herd} objects.
}
\details{
Format of \code{value} for the class replacement methods:

\describe{\item{\verb{bf:}}{A length-one numeric vector containing the
buffer factor of the medium. This value is used to calculate the proton
efflux rate (PER) from the extracellular acidification rate (ECAR).}}

\describe{\item{\verb{cells:}}{A list or data frame containing normalization
data for an experiment. \code{cells} must contain a character vector named \code{well}
and a numeric vector named \code{value}. \code{well} must contain a unique entry for
each well in the experimental plate. \code{value} contains the normalization
factor (\emph{e.g.}, cell number or protein content).}}

\describe{\item{\verb{cf:}}{A length-one numeric vector describing the CO2
correction factor for the cells in the assay. This parameter is used to
separate the mitochondrial and glycolytic contributions to the proton
efflux rate.}}

\describe{\item{\verb{stages:}}{A list or data frame containing vectors
of experimental descriptors for each of the measurement stages. At
minimum, \code{stages} must contain character vectors named \code{measurement} and
\code{stage}. \code{measurement} must contain a unique entry for each measurement
in the experiment. \code{stage} describes the experimental interval (\emph{e.g.},
for a typical mitochondrial stress assay, \code{stage} would be \code{basal},
\code{oligo}, \code{fccp}, \code{rot/ama}). A \code{well} column may also be included if the
injections differ across the plate.}}

\describe{\item{\verb{units:}}{A character vector describing the units of
the normalization factor provided in the \code{value} column of \code{cells}. This
will be used to annotate graphs.}}

\describe{\item{\verb{units:}}{A character vector describing the units of
the normalization factor. This will be used to annotate graphs.}}

\describe{\item{\verb{wells:}}{A list or data frame containing vectors
of experimental descriptors for each of the sample wells. \code{wells} must
contain character vectors named \code{well} and \code{type}. \code{well} must contain
a unique entry for each well in the experimental plate. \code{type} must be
\code{sample} or \code{blank} for each well. Users should consider a \code{group} column
that labels unique experimental groups. Additional descriptors may be
included (\emph{e.g.}, \code{treatment}, \code{time}, \code{condition}, \emph{etc.}). The
experimental data provided will be converted to a data frame. If a
\code{group} vector was not provided, one will be generated from \code{type} or a
combination of the remaining descriptors. Assigning an empty list with
\verb{wells<-} will restore the assignment based on the raw data file.}}
}
