% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\alias{seahorse_levels}
\alias{rates}
\alias{seahorse_summary}
\alias{mst}
\alias{gst}
\alias{atp}
\alias{Seahorse_analysis}
\alias{levels}
\alias{levels,Seahorse-method}
\alias{rates,Seahorse-method}
\title{Accessors for analysis output}
\usage{
levels(x, ...)

\S4method{levels}{Seahorse}(x, blanks = TRUE, outliers = FALSE)

rates(x, ...)

\S4method{rates}{Seahorse}(x, blanks = TRUE, outliers = FALSE, normalize = TRUE)
}
\arguments{
\item{x}{An \strong{R} object. Currently there are methods for
\link{Seahorse} and \linkS4class{Herd} objects.}

\item{...}{Additional arguments passed to other functions.}

\item{blanks}{Include blanks in output?}

\item{outliers}{Include outliers in output?}

\item{normalize}{Divide values by normalization factor?}

\item{object}{An \strong{R} object.}
}
\description{
These functions access calculated values from \code{Seahorse} or
\code{Herd} objects. Collecting \code{Seahorse} objects into a \code{Herd} returns one
average value for each group from each \code{Seahorse} experiment. For
example, \code{summary(Seahorse)} returns summarized values for each stage
from each well, while \code{summary(Herd)} returns summarized values for each
stage from each group. Note that \code{summary} functions automatically
exclude outliers.
}
\section{Functions}{
\itemize{
\item \code{levels()}: The O~2~ and pH measurements used to calculate oxygen
consumption and extracellular acidification rates, respectively.

\item \code{rates()}: The oxygen consumption and extracellular acidification
or proton efflux rates.

}}
\examples{
levels(sheldon)
levels(sheldon, blanks = FALSE, outliers = FALSE)

rates(sheldon)
rates(sheldon, blanks = FALSE, outliers = FALSE)
}
