% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\alias{summarize_rates}
\alias{summarize_mst}
\alias{summarize_gst}
\alias{summarize_atp}
\title{Analysis functions for Seahorse rate data}
\usage{
summarize_rates(x)

summarize_mst(x)

summarize_gst(x)

summarize_atp(x, bf, cf)
}
\arguments{
\item{x}{A data.frame returned by \code{rates()}.}

\item{bf}{The buffer factor of the assay medium.}

\item{cf}{The carbon dioxide correction factor.}
}
\description{
These functions operate on a \code{rates} data.frame from a \code{Seahorse}
object to return summary values from various experimental designs
including mitochondrial stress tests, glycolytic stress tests, and ATP
production rates. These analyses are automatically performed during the
creation of a \code{Seahorse} object, but these functions are exposed to the
user to facilitate finer-grained analyses.
}
\section{Functions}{
\itemize{
\item \code{summarize_rates()}: Calculates summary values for each interval of a
\code{Seahorse} object. For \code{basal}, the last measurement in the interval is
returned. For \code{oligo} and \code{rot/ama}, the lowest OCR and highest ECAR
values are returned. For \code{fccp}, the highest OCR and ECAR values are
returned. For any other interval, the value of the last measurement in
the interval is returned.

\item \code{summarize_mst()}: Calculates the spare respiratory capacity (\code{src}) and
coupling efficiency (\code{coupling}) from a mitochondrial stress test. This
test requires stages labeled \code{basal}, \code{oligo}, \code{fccp}, and \code{rot/ama}.
Both values are expressed as the fold change relative to the basal OCR
after subtracting non-mitochondrial OCR.

\item \code{summarize_gst()}: Calculates the results of a glycolytic stress test. This
test requires stages labeled \code{basal} and \code{oligo}. The \code{gst} value is the
fold increase in ECAR or PER following oligomycin treatment.

\item \code{summarize_atp()}: Calculates the mitochondrial and glycolytic ATP
production rates. This test requires stages labeled \code{basal}, \code{oligo}, and
\code{rot/ama}. In addition, the assay medium buffer factor (\code{bf}) and carbon
dioxide correction factor (\code{cf}) must also be provided.

}}
