% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanks.R
\name{blanks}
\alias{blanks}
\alias{blanks<-}
\alias{blanks,Seahorse-method}
\alias{blanks<-,Seahorse-method}
\title{Blanks accessors for \code{seahorse} S4 objects}
\usage{
\S4method{blanks}{Seahorse}(x)

\S4method{blanks}{Seahorse}(x, action = c("remove", "reset", "replace", "add", "subtract")) <- value
}
\arguments{
\item{x}{A \code{Seahorse} object.}

\item{action}{Indicates how replacement values modify the existing values:\tabular{ll}{
   \strong{action} \tab     \strong{description} \cr
   remove \tab     set all values to \code{NA} \cr
   reset \tab     assign blanks based on the original data file \cr
   replace \tab     completely replace \code{blanks} with new values \cr
   add \tab     add new blank wells \cr
   subtract \tab     remove blank wells \cr
}}

\item{value}{A data frame with two columns named \code{rate} and \code{well}. The
\code{rate} columns should contain either \code{OCR} or \code{ECAR}. The \code{well} column
identifies the blank wells formatted as "A01". Changing blank wells
automatically recalculates OCR and ECAR values. For the "reset" and
"remove" actions, the assigned value is ignored, but use \code{NA}.}
}
\value{
\code{blanks(x)} returns a data frame of wells used as blanks for OCR
and ECAR calculations.
}
\description{
Blank values are essential for accurate calculations of oxygen consumption
(OCR) and extracellular acidification (ECAR) rates. Wells removed from
\code{blanks} are added to the list of experimental outliers.
}
\examples{
blanks(sheldon)
blanks(sheldon, "replace") <- list(OCR = "A01")
blanks(sheldon)
blanks(sheldon, "reset") <- NA
blanks(sheldon)

}
