% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples}
\alias{examples}
\alias{harold}
\alias{horsea}
\alias{sheldon}
\alias{storm}
\alias{herd}
\title{Seahorse data objects}
\format{
\tabular{ll}{
   \strong{name} \tab     \strong{description} \cr
   harold \tab     The seahorse herald from Disney's \emph{The Little Mermaid} \cr
   horsea \tab     A water-type Pokémon that evolves into Seadra \cr
   sheldon \tab     The water-allergic seahorse form Pixar's \emph{Finding Nemo} \cr
   storm \tab     Aquaman's noble steed \cr
   herd \tab     A \code{Herd} object containing the four example data sets \cr
}
}
\description{
These Seahorse objects enable exploration of the functions within the
package.
}
\keyword{datasets}
