% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{findOut}
\alias{findOut}
\alias{findOut,Seahorse-method}
\title{Remove outlying samples}
\usage{
findOut(x, ...)

\S4method{findOut}{Seahorse}(x, blanks = TRUE, outliers = FALSE, ...)
}
\arguments{
\item{x}{An \strong{R} object. Currently there are methods for
\link{Seahorse} and \linkS4class{Herd} objects.}

\item{...}{Reserved for future development}

\item{blanks}{Should \code{findOut} look for outliers in blank samples?}

\item{outliers}{Should \code{findOut} include current outliers when looking for
new outliers?}
}
\value{
Returns a data frame of outlying samples.
}
\description{
Generic function for removing outliers.
}
\section{Methods (by class)}{
\itemize{
\item \code{findOut(Seahorse)}: This method compares OCR and ECAR rates within
experimental groups to identify outlying wells. The function returns a
data frame identifying outlying wells. These wells can then be excluded
from downstream analyses by assigning to the \code{Seahorse} object using
\verb{outliers<-}.

}}
\examples{
outliers(sheldon, "replace") <- findOut(sheldon)
}
