% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cells.R
\name{format_cells}
\alias{format_cells}
\title{Import table of normalization values}
\usage{
format_cells(path, blanks = c("A01", "B04", "C03", "D06"))
}
\arguments{
\item{path}{Path to the data file}

\item{blanks}{A character vector of blank well IDs. If no blanks were
included in the experiment, set to \code{NA}.}
}
\value{
A \code{tbl_df} of normalization data
}
\description{
This function will read in a \code{.csv} file of normalization data.
}
\details{
The file should be organized the same as the Seahorse microplate. The header
row must be \code{row} followed by the column number formatted as \code{x01}. The \code{row}
column should be the letter identification for each row:\tabular{lrrrrrr}{
   row \tab  x01 \tab  x02 \tab  x03 \tab  x04 \tab  x05 \tab   x06 \cr
   A \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \cr
   B \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \cr
   C \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \cr
   D \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \tab  ... \cr
}


Example data may also be found in the \code{extdata} folder of the package source.
The average blank value will be subtracted from the sample values.
}
\examples{
path <- system.file("extdata/counts_1.csv", package = "seahorse", mustWork = TRUE)
format_cells(path, blanks = c("A01", "B04", "C03", "D06"))

}
