% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helpers}
\alias{helpers}
\alias{wells_24}
\alias{stages_mst}
\alias{wells_ex}
\title{Experimental design helpers}
\format{
\tabular{ll}{
   \strong{helper} \tab     \strong{description} \cr
   wells_24 \tab     a character vector of 24-well plate IDs \cr
   wells_ex \tab     example of wells data for \code{Seahorse} construction \cr
   stages_mst \tab     list describing the default design of a mitochondrial stress test \cr
}
}
\description{
These objects should facilitate creation of experimental designs for
\code{Seahorse} object construction.
}
\keyword{datasets}
