% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Herd-class.R
\name{Herd}
\alias{Herd}
\alias{Herd-class}
\title{Herd class constructor}
\usage{
Herd(...)
}
\arguments{
\item{...}{A list of \code{Seahorse} objects.}
}
\description{
\code{Herd} is an S4 container for summarized data from multiple \code{Seahorse}
experiments. These \code{Seahorse} experiments should be completely analyzed
prior to creating a \code{Herd} object, including data normalization and outlier
identification. The combined data automatically exclude blank and outlier
wells from individual experiments. As the data are aggregated, outlying
points are excluded based on median absolute deviation and removed prior to
calculating the average for an experiment.
}
\examples{
Herd(harold, horsea, sheldon, storm)
}
