% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\alias{outliers<-}
\alias{outliers,Seahorse-method}
\alias{outliers<-,Seahorse-method}
\title{Outliers accessors for \code{seahorse} S4 objects}
\usage{
\S4method{outliers}{Seahorse}(x)

\S4method{outliers}{Seahorse}(x, action = c("remove", "reset", "replace", "add", "subtract")) <- value
}
\arguments{
\item{x}{A \code{Seahorse} or \code{Herd} object.}

\item{action}{Indicates how replacement values modify the existing values:\tabular{ll}{
   \strong{action} \tab     \strong{description} \cr
   remove \tab     set all values to \code{NA} \cr
   reset \tab     assign outliers based on \code{findOut} \cr
   replace \tab     completely replace \code{outliers} with new values \cr
   add \tab     add new outlier wells \cr
   subtract \tab     remove outlier wells \cr
}}

\item{value}{A data frame with two columns named \code{rate} and \code{well}. The
\code{rate} columns should contain either \code{OCR} or \code{ECAR}. The \code{well} column
identifies the outlier wells formatted as "A01". For the "reset" and
"remove" actions, the assigned value is ignored, but use \code{NA}.}
}
\value{
\code{outliers(x)} returns a data frame of wells identified as outliers
for OCR and ECAR calculations and plots.
}
\description{
Outlier wells may be excluded from analyses. They are identified using the
\code{findOut} method.
}
\examples{
outliers(sheldon)
outliers(sheldon, "replace") <- findOut(sheldon)
outliers(sheldon)
outliers(sheldon, "reset") <- NA
outliers(sheldon)

}
\seealso{
\link{findOut}
}
