% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,Seahorse,character-method}
\alias{Seahorse_plot}
\title{Plot}
\usage{
plot(x, y, ...)

\S4method{plot}{Seahorse,character}(
  x,
  y = c("cells", "levels", "rates", "summary", "mst", "gst", "atp"),
  group = TRUE,
  blanks = TRUE,
  outliers = TRUE,
  normalize = TRUE,
  type = c("scatter", "bar")
)
}
\arguments{
\item{x}{An \strong{R} object. The \code{seahorse} package describes methods for
\link{Seahorse} and \linkS4class{Herd} objects.}

\item{y}{\describe{
\code{cells} = Plot normalization values. \cr
\code{levels} = Plot levels. \cr
\code{rates} = Plot rates. \cr
\code{summary} = Plot summary data. \cr
\code{mst} = Plot mito stress results. \cr
\code{gst} = Plot glyco stress results. \cr
\code{atp} = Plot ATP rate assay results.}}

\item{...}{Additional arguments passed along to \code{levels()} or \code{rates()},
including \code{blanks} and \code{outliers}.}

\item{group}{Plot summary data for each experimental group or data from
individual wells separately.}

\item{normalize}{Divide values by normalization factor}

\item{type}{Plot ATP rate data as a "scatter" plot or "bar" plot.}
}
\value{
A plot.
}
\description{
Generic plotting function.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot(x = Seahorse, y = character)}: This method plots level or rate data from \code{Seahorse}
objects. The plots can include or exclude blank and outlier wells by
passing these arguments to \code{levels()} and \code{rates()}. Similarly,
non-normalized or normalized rate data may be reviewed.

}}
\examples{
plot(sheldon, "levels")
plot(sheldon, "rates", group = FALSE)
plot(sheldon, "rates", group = TRUE, blanks = TRUE)
plot(sheldon, "rates", group = TRUE, blanks = FALSE)

}
\seealso{
\code{\link[=rates]{rates()}}, \code{\link[=levels]{levels()}}
}
