% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seahorse-package.R
\docType{package}
\name{seahorse-package}
\alias{seahorse}
\alias{seahorse-package}
\title{seahorse: Tools for Analyzing Seahorse Bioenergetic Data}
\description{
This package provides tools to import Seahorse data from an exported Excel file; calculate OCR, ECAR, and PER from raw data; deal with outlying measurements in both background and experimental wells; calculate various experimental parameters; generate summary summary statistics; and aggregate multiple experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/seahorse}
  \item Report bugs at \url{https://github.com/wmoldham/seahorse/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
