% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seahorse-class.R
\name{Seahorse}
\alias{Seahorse}
\title{Seahorse class constructor}
\usage{
Seahorse(
  path,
  wells = list(),
  stages = list(),
  cells = list(),
  units = NA_character_,
  bf = NA_real_,
  cf = NA_real_
)
}
\arguments{
\item{path}{Character vector for the Seahorse \code{xlsx} file path.}

\item{wells}{An optional list or data frame containing vectors of
experimental descriptors for each of the sample wells. At minimum,
\code{wells} must contain character vectors named \code{well} and \code{type}. \code{well}
must contain a unique entry for each well in the experimental plate.
\code{type} must be \code{sample} or \code{blank} for each well. Users should consider
a \code{group} column that labels unique experimental groups. It is probably
worthwhile to convert this column to a factor based on the preferred
ordering of experimental groups. Additional descriptors may be included
(\emph{e.g.}, \code{treatment}, \code{time}, \code{condition}). If a \code{group} vector was not
provided, one will be generated from \code{type} or a combination of the
remaining descriptors.}

\item{stages}{An optional list or data frame containing vectors of
experimental descriptors for each of the measurement stages. \code{stages}
must contain character vectors named \code{measurement} and \code{stage}.
\code{measurement} must contain a unique entry for each measurement in the
experiment. \code{stage} describes the experimental interval (\emph{e.g.}, for a
typical mitochondrial stress assay, \code{stage} would be \code{basal}, \code{oligo},
\code{fccp}, or \code{rot/ama}). A \code{well} column may also be included if the
injections differ among treatment wells.}

\item{cells}{An optional list or data frame containing vectors of
normalization data for an experiment. If provided, \code{cells} must contain
a character vector named \code{well} and a numeric vector named \code{value}.
\code{well} must contain a unique entry for each well in the experimental
plate. \code{value} contains the normalization factor (\emph{e.g.}, cell number or
protein content).}

\item{units}{An optional character vector describing the units of the
normalization factor provided in the \code{value}column of \code{cells}. This will
be used to annotate graphs.}

\item{bf}{An optional length-one numeric vector containing the buffer factor
of the medium. This value is used to calculate the proton efflux rate
(PER) from the extracellular acidification rate (ECAR).}

\item{cf}{An optional length-one numeric vector describing the carbon dioxide
correction factor for the cells in the assay. This parameter is used to
separate the mitochondrial and glycolytic contributions to the proton
efflux rate.}
}
\description{
\code{Seahorse} is an S4 container used to store the calibration values and raw
data from a single Seahorse assay; to calculate oxygen consumption,
extracellular acidification, and proton efflux rates from the raw data; and
to facilitate outlier detection, data analysis, and visualization.
}
\details{
During creation of the Seahore object, raw fluorescence readings
are converted to O2 and pH measurements from which oxygen consumptions
and extracellular acidification rates are calculated. Outlying wells are
identified and removed.
}
\examples{
path <- system.file("extdata/raw_1.xlsx", package = "seahorse", mustWork = TRUE)
well <- wells_24
type <- c("blank", rep("sample", 8), "blank", rep("sample", 4), "blank", rep("sample", 8), "blank")
x <- rep(LETTERS[1:5], each = 4)
group <- replace(type, which(type == "sample"), x)
wells <- list(well = well, type = type, group = group)
measurement <- 1:12
stage <- rep(LETTERS[1:4], each = 3)
stages <- list(measurement = measurement, stage = stage)
cells <- list(well = well, value = 20000)
Seahorse(path, wells, stages, cells, units = "cell", bf = 2.4, cdcf = 0.41)

}
\seealso{
\link{helpers}, \link{format_cells}, \link{accessors}, \link{blanks}, \link{outliers}, \link{plot}
}
